TransmissionMatrix tests: fixed TF input tests and moved other related input tests to same file
